requirements.txt
pyyaml
pymongo
selenium
python-telegram-bot
telethon
dnspython
python3-dateutil
tabulate